logger/requirements.txt
redis==3.5.3
pyserial==3.5
